Created a LAPVQ object, which has an optional argument called epsilon in its constructor.
We let the epsilon be embedded into the object, rather than using it as an argument for its encoding function, so we can call the forward methods without changing anything.

This means, that whenever we wish to change epsilon, we have to use "set_epsilon".

Changed the faceshq_transformer.yaml to use the LAPVQ script.

Added set_epsilon in cond_transformer.py Net2NetTransformer.